+++ directdemod/constants.py
@@ -22,6 +22,9 @@ NOAA_DETECTMAXCHANGE = 5
 NOAA_DETECTCONSSYNCSNUM = 10
 NOAA_SATS = {137620000:"NOAA 15", 137100000:"NOAA 19", 137912500:"NOAA 18"}
 
+## Merger settings
+RESOLUTION = 1000
+COLOR = "black"
 
 ###### Do not change these
