--- directdemod/merger/config.py
@@ -1,4 +1,0 @@
-### Configuration file for ImageMerger class
-
-RESOLUTION=1000
-COLOR="black"
