+++ directdemod/misc.py
